backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]>=3.1.0
python-decouple==3.8
Pillow==12.1.0
